code cleanup refactor
