+++ tasks/deposit-extra.js
@@ -1,28 +1,7 @@
-const fs = require("fs").promises;
-const path = require("path");
 const addressesPath = "./scripts/migration/addresses";
+const { readJsonFiles } = require("./utils");
 
-async function readJsonFiles(dirPath) {
-  const vaults = new Map();
-  try {
-    const files = await fs.readdir(dirPath);
-    const jsonFiles = files.filter((file) => path.extname(file).toLowerCase() === ".json");
-
-    for (const file of jsonFiles) {
-      const filePath = path.join(dirPath, file);
-      const fileContent = await fs.readFile(filePath, "utf8");
-      const jsonData = JSON.parse(fileContent);
-      const modifiedName = file.replace("mainnet_", "").replace(".json", "");
-      vaults.set(modifiedName, jsonData);
-    }
-  } catch (error) {
-    console.error("Error reading JSON files:", error);
-  }
-
-  return vaults;
-}
-
-task("deposit-extra", "TODO")
+task("deposit-extra", "Deposits extra assets straight away to EigenLayer")
   .addParam("vault", "The name of the vault")
   .setAction(async (taskArgs) => {
     const inputVaultName = taskArgs["vault"];
