+++ tasks/get-free-balances.js
@@ -1,28 +1,7 @@
-const fs = require("fs").promises;
-const path = require("path");
 const addressesPath = "./scripts/migration/addresses";
+const { readJsonFiles } = require("./utils");
 
-async function readJsonFiles(dirPath) {
-  const vaults = new Map();
-  try {
-    const files = await fs.readdir(dirPath);
-    const jsonFiles = files.filter((file) => path.extname(file).toLowerCase() === ".json");
-
-    for (const file of jsonFiles) {
-      const filePath = path.join(dirPath, file);
-      const fileContent = await fs.readFile(filePath, "utf8");
-      const jsonData = JSON.parse(fileContent);
-      const modifiedName = file.replace("mainnet_", "").replace(".json", "");
-      vaults.set(modifiedName, jsonData);
-    }
-  } catch (error) {
-    console.error("Error reading JSON files:", error);
-  }
-
-  return vaults;
-}
-
-task("get-free-balance", "It returns the current state of Inception").setAction(async (taskArgs) => {
+task("get-free-balances", "It returns the current state of Inception").setAction(async (taskArgs) => {
   const vaults = await readJsonFiles(addressesPath);
   for (const [vaultName, vaultData] of vaults) {
     await getState(vaultName, vaultData.iVaultAddress);
